Request access to nova and nova_api databases
The DataMover API requires access to the nova and nova_api databases
(not the dmapi database).

Update handler to request access to the correct shared database and
drop provision of unit private address - the interface does the
right thing with regards to using network space binding of the
relation endpoint on the local charm unit.