*WIP* Do not merge - First pass at overlay bundle and test fixture
Distill the Trilio charm bits out of the test bundle, place them
into a separate overlay bundle as an example. Users will still
need to inject their own charm config values into the overlay
bundle before deployment. This allows devs/users to simply do
something like:

    juju deploy openstack-base --overlay myoverlay.yaml

Add a swift test fixture as a separate overlay bundle, to create
a stand-alone swift s3 target, with a separate keystone and db.
This is to simulate a backup target storage cluster. Because it
has its own separate keystone and db, it is not related in any
way to the openstack or Trilio charms which are under test.