Update endpoint map configuration
Unfortunately you really can't update the ServiceNetMap and EndpointMap
nicely.  You can however overide some parts of it with
EndpointMapOverride.  The EndpointMapOverride is useful when wanting to
switch ports or protocols, but it does not work with the composable
networks. This change creates a script that can generate the correct
endpoint map depending on which version is being used and included
environment files to override ports/protocols when enabling tls. This
change includes a queens and train version of the endpoint maps.