Update tls files
So the EndpointMapOverride modifies the generated EndpointMap and not
the source input to it. So in order to change ports and protocols for
when enabling tls, you must specify the complete EndpointMap. In
tripleo, that's handled by our provided tls environment files. Since we
do not have a way to append to these, they should be shipped with Trilio
specific information them.  Documentation for Trilio should be updated
to indicate if a customer wants to enable TLS, they need to use the
trilio specific environment files instead of the tripleo version or the
information needs to be appended to another file if another
customization has been done by the end user. That being said, these
files need to be kept insync with the shipped ones.  This change has a
queens and a train version for each of these files.